Add binding data from sessionStorage
Task: AFORM-3713
