+++ src/@optimizely/forms-sdk/src/helpers/index.ts
@@ -3,3 +3,4 @@ export * from "./urlHelper";
 export * from "./elementHelper";
 export * from "./utils";
 export * from "./fileHelper";
+export * from "./initFormState";
