+++ src/@optimizely/forms-sdk/src/helpers/utils.ts
@@ -3,7 +3,7 @@
  * @param value A string value to check
  * @returns True or false
  */
-export function isNullOrEmpty(value: string | undefined): boolean{
+export function isNullOrEmpty(value?: string): boolean{
     return isNull(value) || value === "";
 }
